+++ packages/components/src/components/keyboard/keyboard-alphabetic-layouts.ts
@@ -57,4 +57,14 @@ export const spanishAlphabetic = {
     ]
 };
 
+export const dutchAlphabetic = {
+    default: [
+        '1 2 3 4 5 6 7 8 9 0 {bksp}',
+        'q w e r t y u i o p',
+        'a s d f g h j k l',
+        'z x c v b n m . - &',
+        '{space}'
+    ]
+};
+
 export const defaultAlphabetic = unitedStatesAlphabetic;