+++ packages/components/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.22.0] - 2024-09-19
+
+### Added
+
+- `cleared` event emitter to `mi-dropdown` component. It notifies when dropdown search input is cleared.
+
+## [13.21.0] - 2024-09-18
+
+### Added
+
+- `clearFilter` function for `mi-dropdown`.
+
+## [13.20.3] - 2024-09-18
+
+### Added
+
+- Added new property to the `mi-dropdown` component called `placeholder`.
+
 ## [13.20.2] - 2024-07-17
 
 ### Fixed